GCBI: enable running on Modules without root
This commit enables running the GCBI pass on Modules without a `root`
Function. This is necessary in decompilation, where we handle Modules
with only isolated functions.

If the pass runs on a Module without `root`, information about
`BlockType`s cannot be computed, but everything else (SP, PC, and other
architectural info) is properly recovered for use in decompilation.