OpaqueFunctionsPool: rework record mechanism
This commit slightly tweaks the operation of the pools when recording a
new function.
Specifically, now all the initialization of the pools from
`llvm::Module`s happens through initializer methods. This enables the
initialization method to skip some sanity checks on the consistency,
using a newly provided `recordUnchecked` method.
This new `recordUnchecked` method is kept private, because skipping
checks might be dangerous, and should be done with care.
Only the old checked `record` method is available as a public API.