MakeModelCastPass: postpone arguments computation
This commits reworks MakeModelCastPass so that the computation the of
arguments for the invocation of a ModelCast are now computed later: not
when we decided that a cast must be emitted, but right before we emit
the call.

This simplifies the code and lays the groundwork for relaxing some
assertions later.