DecompileFunction: emit comment on invalid casts
This commit ensures that, when the LLVM-based C code generation emits a
cast that cannot compile in C, it explicitly marks it with a block
comment.

This is necessary because the LLVM-based C backend is deprecated in
favor of the clift-based decompilation pipeline, so we're not actively
fixing bugs anymore, and it currently cannot guarantee that such invalid
casts are never emitted.